+++ app/scripts/services/searches.ts
@@ -1,10 +1,10 @@
 /** @format */
 import { mergeCqpExprs, parse, stringify } from "@/cqp_parser/cqp"
-import { updateSearchHistory } from "@/history"
 import { RootScope } from "@/root-scope.types"
+import { SearchHistoryService } from "@/services/search-history"
 import { LocationService } from "@/urlparams"
-import { unregescape } from "@/util"
-import angular, { IDeferred, IQService } from "angular"
+import angular, { IDeferred, IQService, ITimeoutService } from "angular"
+import "@/services/search-history"
 
 export type SearchesService = {
     activeSearch: {
@@ -16,6 +16,7 @@ export type SearchesService = {
     langDef: IDeferred<never>
     kwicSearch: (cqp: string) => void
     getCqpExpr: () => string
+    triggerSearch: () => void
 }
 
 /**
@@ -28,9 +29,17 @@ export type SearchesService = {
  */
 angular.module("korpApp").factory("searches", [
     "$location",
-    "$rootScope",
     "$q",
-    function ($location: LocationService, $rootScope: RootScope, $q: IQService): SearchesService {
+    "$rootScope",
+    "$timeout",
+    "searchHistory",
+    function (
+        $location: LocationService,
+        $q: IQService,
+        $rootScope: RootScope,
+        $timeout: ITimeoutService,
+        searchHistory: SearchHistoryService
+    ): SearchesService {
         const searches: SearchesService = {
             activeSearch: null,
             langDef: $q.defer(),
@@ -45,6 +54,13 @@ angular.module("korpApp").factory("searches", [
                 if (this.activeSearch.type === "word" || this.activeSearch.type === "lemgram")
                     return $rootScope.simpleCQP || ""
                 return this.activeSearch.val
+            },
+
+            triggerSearch(): void {
+                // Unset and set in next tick, to trigger our watcher
+                const search = $location.search().search
+                $location.search("search", null)
+                $timeout(() => $location.search("search", search))
             },
         }
 
@@ -61,8 +77,7 @@ angular.module("korpApp").factory("searches", [
                 // Store new query in search history
                 // For Extended search, `value` is empty (then the CQP is instead in the `cqp` URL param)
                 if (value) {
-                    const historyValue = type === "lemgram" ? unregescape(value) : value
-                    updateSearchHistory(historyValue, $location.absUrl())
+                    searchHistory.addItem($location.search())
                 }
                 $q.all([searches.langDef.promise, $rootScope.globalFilterDef.promise]).then(function () {
                     if (type === "cqp") {